Toolkit: Material blueprint asset compiler: Implemented sampler state
Former-commit-id: 7e18cbd4eb2a17499a778510f2e43cca61ef5cd1
Former-commit-id: 8fd5bc7f756d8e0044bede1f931e8690ab162e66

+++ Renderer/RendererToolkit/src/AssetCompiler/CompositorAssetCompiler.cpp
@@ -224,7 +224,7 @@ namespace RendererToolkit
 											{
 												for (size_t i = 0; i < 4; ++i)
 												{
-													passClear.color[i] = static_cast<float>(std::atof(stringTokenizer[i].c_str()));
+													passClear.color[i] = std::stof(stringTokenizer[i].c_str());
 												}
 											}
 											else
